<refact> : <Project API 작성>
-Project update, read 작성(진행중)

+++ src/main/java/com/example/tripvision/project/application/ProjectServiceImpl.java
@@ -43,8 +43,9 @@ public class ProjectServiceImpl implements ProjectService {
 
 
 	@Transactional(readOnly = true)
-	public Project findProject(Long id) {
-		return projectRepository.findById(id).orElseThrow(() -> new RuntimeException("해당 프로젝트가 없습니다."));
+	public Project findProject(Long teamId) {
+		Team team = Team.builder().id(teamId).build();
+		return projectRepository.findByTeam(team).orElseThrow(() -> new RuntimeException("해당 프로젝트가 없습니다."));
 	}
 
 	@Transactional
@@ -60,15 +61,29 @@ public class ProjectServiceImpl implements ProjectService {
 
 	@Transactional
 	public Project updateProject(Project project) {
-		Project newProject = projectRepository.findById(project.getId())
-			.orElseThrow(() -> new RuntimeException("good"));
-		newProject.update(project);
-		return newProject;
+
+		Team team = teamRepository.findTeamLazy(project.getTeam().getId())
+			.orElseThrow(() -> new RuntimeException("팀을 찾지 못했습니다."));
+
+		Project updateProject = team.getProject();
+		updateProject.update(project);
+
+		return updateProject;
 	}
 
 	@Transactional
-	public void deleteProject(Long projectId) {
+	public void deleteProject(Long teamId, Long projectId) {
+		Project project = projectRepository.findByIdLazy(projectId)
+				.orElseThrow(() -> new RuntimeException("프로젝트 못찾음"));
+
+		if (!project.isMatchTeamId(teamId)){
+			throw new RuntimeException("프로젝트의 팀 ID가 일치하지 않습니다.");
+		}
+
+		project.getBudget().deleteProject();
+		project.deleteProject(project.getTeam());
 		projectRepository.deleteById(projectId);
+
 		// 단방향 이면, 따로 연관관계를 삭제해줍시다.
 		// 양방향 이면, Cascade 를 이용해서 삭제 해줍시다.
 	}
